retriever/dense.py
"""
Dense Retriever using sentence embeddings.

This module implements a retriever that finds relevant documents based on the
semantic similarity of their embeddings. It uses a pre-trained sentence-transformer
model to generate embeddings and a vector index (like FAISS) for efficient search.
"""
import faiss
from sentence_transformers import SentenceTransformer
from typing import List
from ..utils.schema import DocumentChunk
from ..utils.logging import get_logger

logger = get_logger(__name__)

class DenseRetriever:
    """
    A retriever that uses dense vector embeddings to find similar documents.
    """
    def __init__(self, model_name: str, index_path: str):
        """
        Initializes the DenseRetriever.

        Args:
            model_name (str): The name of the sentence-transformer model to use.
            index_path (str): The path to the pre-built FAISS index.
        """
        # TODO: Add a method to build the index from a corpus of documents if it doesn't exist.
        # This would involve loading documents, chunking them, encoding them, and saving the index.
        logger.info(f"Initializing dense retriever with model: {model_name}")
        try:
            self.model = SentenceTransformer(model_name)
            self.index = faiss.read_index(index_path)
            # TODO: The mapping from index ID to document content needs to be loaded.
            # This is a critical piece that's missing in this simple implementation.
            # For now, we'll assume a placeholder `id_to_doc` mapping.
            self.id_to_doc = self._load_document_map()
            logger.info(f"FAISS index loaded successfully from {index_path}")
        except Exception as e:
            logger.error(f"Failed to initialize DenseRetriever: {e}")
            self.model = None
            self.index = None
            self.id_to_doc = {}

    def _load_document_map(self) -> dict:
        """
        Placeholder for loading the mapping from FAISS index IDs to document chunks.
        In a real implementation, this would load a pickle file, a JSON file, or a database table.
        """
        # TODO: Implement the actual loading of the document map.
        logger.warning("Using placeholder document map. Implement `_load_document_map`.")
        return {
            0: {"source": "doc1.txt", "content": "Quantum computing is a new paradigm."},
            1: {"source": "doc2.txt", "content": "Error correction is key in quantum systems."},
        }

    def retrieve(self, query: str, top_k: int = 5) -> List[DocumentChunk]:
        """
        Retrieves the top_k most relevant document chunks for a given query.

        Args:
            query (str): The user's query.
            top_k (int): The number of documents to retrieve.

        Returns:
            List[DocumentChunk]: A list of relevant document chunks.
        """
        if not self.model or not self.index:
            logger.error("Dense retriever is not initialized. Cannot retrieve.")
            return []

        logger.debug(f"Retrieving top {top_k} documents for query: '{query}'")
        try:
            query_embedding = self.model.encode([query])
            distances, indices = self.index.search(query_embedding, top_k)

            results = []
            for i in range(len(indices[0])):
                doc_id = indices[0][i]
                score = 1 - distances[0][i] # Convert distance to similarity score
                if doc_id in self.id_to_doc:
                    doc = self.id_to_doc[doc_id]
                    results.append(
                        DocumentChunk(
                            source=doc.get("source", "unknown"),
                            content=doc.get("content", ""),
                            score=score,
                            metadata={"retriever": "dense"}
                        )
                    )
            logger.info(f"Retrieved {len(results)} documents with dense retriever.")
            return results
        except Exception as e:
            logger.error(f"Error during dense retrieval: {e}")
            return []

# TODO: Add other vector database clients like Pinecone or Weaviate as alternative backends.
# class PineconeRetriever:
#     def __init__(self, api_key: str, environment: str, index_name: str):
#         ...
#     def retrieve(self, query: str, top_k: int = 5) -> List[DocumentChunk]:
#         ...
