requirements.txt
openai
faiss-cpu
rank-bm25
elasticsearch
pydantic
requests
rich
pytest
pyyaml
sentence-transformers
cross-encoder
